refactor(issue-dialog): extract status badges, jira fields, SubmitFieldsDialog, and shared FieldRow
Decompose the monolithic dialog hosts so each piece lives in a single
responsibility module. No behavior change — file moves only.

IssueCreateModal.tsx: 1349 -> 589 lines.
IssueListTab.tsx: 1305 -> 243 lines.

New modules:
- src/sidepanel/components/FieldRow.tsx
  Lifted out of IssueCreateModal so the 4 platform fields can import it
  without coupling to the host. github/linear/notion fields updated.

- src/sidepanel/tabs/statusBadges/{constants,PlatformChip,GithubStatusBadge,
  JiraStatusBadge,LinearStatusBadge,NotionStatusBadge,SubmittedBadge}.tsx
  4 platform status badges + the SubmittedBadge wrapper that dispatches
  by platform, the inline color constants, and PlatformChip.

- src/sidepanel/tabs/IssueRow.tsx
  Issue list row with the submitted-vs-draft branching, delete dialog
  for drafts, and SubmittedBadge mount for submitted entries.

- src/sidepanel/tabs/issueListUtils.ts
  Pure helpers: StatusFilter type, isRefreshable, resolveNotionPageId,
  parseGithubIssueNumber, parseGithubIssueUrl, resolveGithubCoords,
  matchesQuery, matchesStatus, formatIssueKey, issueTimestamp,
  dateLabel, formatDate, groupByDate. Tests renamed to match
  (issueListFilters.test.ts -> issueListUtils.test.ts, import path
  updated to ../issueListUtils).

- src/sidepanel/tabs/jiraFields/{JiraIssueFields,IssueTypeField,
  PriorityField,AssigneeField,EpicField,FieldCombobox,
  useDebouncedSearch,useJiraConfig}.tsx
  Restores the 4-platform symmetry — jira now lives in its own
  directory like github/linear/notion. JiraFieldsBlock renamed to
  JiraIssueFields to match the sibling pattern.

- src/sidepanel/tabs/SubmitFieldsDialog.tsx
  Lifted out of IssueCreateModal so DraftDetailDialog imports it from
  a peer location instead of reaching into the other host's internals.
  Both hosts now import { SubmitFieldsDialog } from "./SubmitFieldsDialog".

Verification:
- pnpm typecheck — passes.
- pnpm test — 854 / 854 pass (including 42 cases in the renamed
  issueListUtils.test.ts).
