+++ src/sidepanel/components/FieldRow.tsx
@@ -1,0 +1,21 @@
+import type { ReactNode } from "react";
+
+export function FieldRow({
+  label,
+  required,
+  children,
+}: {
+  label: string;
+  required?: boolean;
+  children: ReactNode;
+}) {
+  return (
+    <div className="grid gap-1.5">
+      <label className="text-xs text-muted-foreground">
+        {label}
+        {required ? <span className="ml-0.5 text-destructive">*</span> : null}
+      </label>
+      {children}
+    </div>
+  );
+}
