+++ src/sidepanel/tabs/IssueCreateModal.tsx
@@ -1,60 +1,16 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import {
-  Check,
-  ChevronsUpDown,
-  Loader2,
-  X,
-} from "lucide-react";
-import {
-  SiGithub,
-  SiJirasoftware,
-  SiLinear,
-  SiNotion,
-} from "@icons-pack/react-simple-icons";
-import { toast } from "sonner";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import { Button } from "@/components/ui/button";
-import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from "@/components/ui/command";
-import {
-  Dialog,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { useT } from "@/i18n";
-import { cn } from "@/lib/utils";
 import { blobToDataUrl, pruneOrphanInlineImages } from "@/store/blob-db";
 import { useSettingsUiStore } from "@/store/settings-ui-store";
-import { useEditorStore, type EditorIssueFields } from "@/store/editor-store";
+import { useEditorStore } from "@/store/editor-store";
 import { useIssuesStore } from "@/store/issues-store";
 import {
   connectedPlatforms,
-  isJiraAccountComplete,
-  isLinearAccountComplete,
-  isNotionAccountComplete,
   jiraSiteId,
   pickInitialPlatform,
   useSettingsStore,
 } from "@/store/settings-store";
-import type {
-  JiraIssueSummary,
-  JiraIssueType,
-  JiraPriority,
-  JiraUser,
-} from "@/types/jira";
 import type { PlatformId } from "@/types/platform";
 import { sendBg, type JiraSubmitResult } from "@/types/messages";
 import { buildStyleDiff } from "@/sidepanel/components/StyleChangesTable";
@@ -74,26 +30,20 @@ import { submitToNotion, type NotionFileInput } from "@/sidepanel/lib/submitToNo
 import { recordingFilename } from "@/sidepanel/lib/video-mime";
 import { extractInlineRefs, resolveInlineImagesForSections, type InlineImageInput } from "@/sidepanel/lib/resolveInlineImages";
 import {
-  GithubIssueFields,
   initialGhFields,
   type GithubIssueFieldsValue,
 } from "./githubFields/GithubIssueFields";
 import {
-  LinearIssueFields,
   initialLinearFields,
   type LinearIssueFieldsValue,
 } from "./linearFields/LinearIssueFields";
 import {
-  NotionIssueFields,
   initialNotionFields,
   type NotionIssueFieldsValue,
 } from "./notionFields/NotionIssueFields";
 import type { NotionDatabaseSchema } from "@/types/notion";
 import { extractNotionPageId } from "@/lib/notion-page-id";
-
-type SubmitState =
-  | { status: "idle" }
-  | { status: "submitting" };
+import { SubmitFieldsDialog } from "./SubmitFieldsDialog";
 
 export function IssueCreateModal() {
   const t = useT();
@@ -635,715 +585,5 @@ export function IssueCreateModal() {
         onSubmit={handleSubmit}
       />
     </>
-  );
-}
-
-export interface SubmitFieldsDialogProps {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  title?: string;
-  platform: PlatformId;
-  setPlatform: (p: PlatformId) => void;
-  availablePlatforms: PlatformId[];
-  jiraFields: EditorIssueFields;
-  setJiraFields: (patch: Partial<EditorIssueFields>) => void;
-  ghFields: GithubIssueFieldsValue;
-  setGhFields: (patch: Partial<GithubIssueFieldsValue>) => void;
-  linearFields: LinearIssueFieldsValue;
-  setLinearFields: (patch: Partial<LinearIssueFieldsValue>) => void;
-  notionFields: NotionIssueFieldsValue;
-  setNotionFields: (patch: Partial<NotionIssueFieldsValue>) => void;
-  onNotionSchemaResolved: (schema: NotionDatabaseSchema | null) => void;
-  onSubmit: (platform: PlatformId) => Promise<NormalizedSubmitResult>;
-  onSuccess?: (result: NormalizedSubmitResult) => void;
-}
-
-export function SubmitFieldsDialog(props: SubmitFieldsDialogProps) {
-  const {
-    open,
-    onOpenChange,
-    title,
-    platform,
-    setPlatform,
-    availablePlatforms,
-    jiraFields,
-    setJiraFields,
-    ghFields,
-    setGhFields,
-    linearFields,
-    setLinearFields,
-    notionFields,
-    setNotionFields,
-    onNotionSchemaResolved,
-    onSubmit,
-    onSuccess,
-  } = props;
-  const t = useT();
-  const jiraAccount = useSettingsStore((s) => s.accounts.jira);
-  const ghAccount = useSettingsStore((s) => s.accounts.github);
-  const linearAccount = useSettingsStore((s) => s.accounts.linear);
-  const notionAccount = useSettingsStore((s) => s.accounts.notion);
-  const [submit, setSubmit] = useState<SubmitState>({ status: "idle" });
-
-  useEffect(() => {
-    if (!open) setSubmit({ status: "idle" });
-  }, [open]);
-
-  const jiraConfigured = isJiraAccountComplete(jiraAccount);
-  const ghConfigured = !!ghAccount;
-  const linearConfigured = isLinearAccountComplete(linearAccount);
-  const notionConfigured = isNotionAccountComplete(notionAccount);
-  const platformConfigured =
-    platform === "jira"
-      ? jiraConfigured
-      : platform === "github"
-        ? ghConfigured
-        : platform === "linear"
-          ? linearConfigured
-          : notionConfigured;
-
-  const canSubmit =
-    submit.status !== "submitting" &&
-    platformConfigured &&
-    (platform === "jira"
-      ? !!jiraFields.issueTypeId
-      : platform === "github"
-        ? !!ghFields.owner && !!ghFields.repo
-        : platform === "linear"
-          ? !!linearFields.teamId
-          : !!notionFields.databaseId);
-
-  async function handleSubmit() {
-    if (!canSubmit) return;
-    setSubmit({ status: "submitting" });
-    try {
-      const result = await onSubmit(platform);
-      onOpenChange(false);
-      onSuccess?.(result);
-    } catch (err) {
-      toast.error(err instanceof Error ? err.message : String(err));
-      setSubmit({ status: "idle" });
-    }
-  }
-
-  function handleOpenChange(next: boolean) {
-    if (submit.status === "submitting") return;
-    onOpenChange(next);
-  }
-
-  const showTabs = availablePlatforms.length > 1;
-
-  return (
-    <Dialog open={open} onOpenChange={handleOpenChange}>
-      <DialogContent className="w-[80vw] max-w-[80vw] gap-5 rounded-3xl p-6 sm:rounded-3xl">
-        <DialogHeader>
-          <DialogTitle className="text-xl">{title ?? t("issue.submit")}</DialogTitle>
-        </DialogHeader>
-
-        {showTabs ? (
-          <Tabs value={platform} onValueChange={(v) => setPlatform(v as PlatformId)}>
-            <TabsList className={cn(
-              "grid h-9 w-full",
-              availablePlatforms.length === 4
-                ? "grid-cols-4"
-                : availablePlatforms.length === 3
-                  ? "grid-cols-3"
-                  : "grid-cols-2",
-            )}>
-              {availablePlatforms.includes("jira") && (
-                <TabsTrigger value="jira" className="gap-1.5">
-                  <SiJirasoftware className="h-3.5 w-3.5" color="default" />
-                  {t("platform.tab.jira")}
-                </TabsTrigger>
-              )}
-              {availablePlatforms.includes("github") && (
-                <TabsTrigger value="github" className="gap-1.5">
-                  <SiGithub className="h-3.5 w-3.5 dark:invert" color="default" />
-                  {t("platform.tab.github")}
-                </TabsTrigger>
-              )}
-              {availablePlatforms.includes("linear") && (
-                <TabsTrigger value="linear" className="gap-1.5">
-                  <SiLinear className="h-3.5 w-3.5" color="default" />
-                  {t("platform.tab.linear")}
-                </TabsTrigger>
-              )}
-              {availablePlatforms.includes("notion") && (
-                <TabsTrigger value="notion" className="gap-1.5">
-                  <SiNotion className="h-3.5 w-3.5 dark:invert" color="default" />
-                  {t("platform.tab.notion")}
-                </TabsTrigger>
-              )}
-            </TabsList>
-          </Tabs>
-        ) : null}
-
-        {platform === "jira" ? (
-          jiraConfigured ? (
-            <JiraFieldsBlock fields={jiraFields} onChange={setJiraFields} />
-          ) : null
-        ) : platform === "github" ? (
-          ghConfigured ? (
-            <GithubIssueFields value={ghFields} onChange={setGhFields} />
-          ) : null
-        ) : platform === "linear" ? (
-          linearConfigured ? (
-            <LinearIssueFields value={linearFields} onChange={setLinearFields} />
-          ) : null
-        ) : notionConfigured ? (
-          <NotionIssueFields
-            value={notionFields}
-            onChange={setNotionFields}
-            onSchemaResolved={onNotionSchemaResolved}
-          />
-        ) : null}
-
-        <DialogFooter className="flex-row justify-end">
-          <Button
-            variant="outline"
-            onClick={() => handleOpenChange(false)}
-            disabled={submit.status === "submitting"}
-          >
-            {t("common.close")}
-          </Button>
-          <Button
-            onClick={() => void handleSubmit()}
-            disabled={!canSubmit}
-            className="relative"
-          >
-            {submit.status === "submitting" && (
-              <span className="absolute inset-0 flex items-center justify-center">
-                <Loader2 className="h-4 w-4 animate-spin" />
-              </span>
-            )}
-            <span className={submit.status === "submitting" ? "opacity-0" : undefined}>
-              {t("common.submit")}
-            </span>
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-function JiraFieldsBlock({
-  fields,
-  onChange,
-}: {
-  fields: EditorIssueFields;
-  onChange: (patch: Partial<EditorIssueFields>) => void;
-}) {
-  const t = useT();
-  return (
-    <div className="flex flex-col gap-4">
-      <FieldRow label={t("create.issueType")} required>
-        <IssueTypeField
-          value={fields.issueTypeId}
-          onChange={(id) => onChange({ issueTypeId: id })}
-        />
-      </FieldRow>
-      <FieldRow label={t("create.assignee")}>
-        <AssigneeField
-          value={fields.assigneeId}
-          fallbackLabel={fields.assigneeName}
-          onChange={(id, name) => onChange({ assigneeId: id, assigneeName: name })}
-        />
-      </FieldRow>
-      <FieldRow label={t("create.priority")}>
-        <PriorityField
-          value={fields.priorityId}
-          fallbackLabel={fields.priorityName}
-          onChange={(id, name) => onChange({ priorityId: id, priorityName: name })}
-        />
-      </FieldRow>
-      <FieldRow label={t("create.parentEpic")}>
-        <EpicField
-          value={fields.parentKey}
-          fallbackLabel={fields.parentLabel}
-          onChange={(key, label) => onChange({ parentKey: key, parentLabel: label })}
-          hierarchyLevels={[1]}
-        />
-      </FieldRow>
-      <FieldRow label={t("create.linkedIssue")}>
-        <EpicField
-          value={fields.relatesKey}
-          fallbackLabel={fields.relatesLabel}
-          onChange={(key, label) => onChange({ relatesKey: key, relatesLabel: label })}
-        />
-      </FieldRow>
-    </div>
-  );
-}
-
-export function FieldRow({
-  label,
-  required,
-  children,
-}: {
-  label: string;
-  required?: boolean;
-  children: React.ReactNode;
-}) {
-  return (
-    <div className="grid gap-1.5">
-      <label className="text-xs text-muted-foreground">
-        {label}
-        {required ? <span className="ml-0.5 text-destructive">*</span> : null}
-      </label>
-      {children}
-    </div>
-  );
-}
-
-function useJiraConfig(): { projectKey: string } | null {
-  const jiraAccount = useSettingsStore((s) => s.accounts.jira);
-  return useMemo(() => {
-    if (!jiraAccount?.projectKey || !jiraAccount.auth) return null;
-    return { projectKey: jiraAccount.projectKey };
-  }, [jiraAccount?.auth, jiraAccount?.projectKey]);
-}
-
-function useDebouncedSearch<T>(
-  fetchFn: (query: string) => Promise<T[]>,
-  delay = 300,
-) {
-  const [items, setItems] = useState<T[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const seqRef = useRef(0);
-
-  const search = useCallback(
-    (query: string) => {
-      const seq = ++seqRef.current;
-      setLoading(true);
-      setError(null);
-      const timer = window.setTimeout(() => {
-        fetchFn(query)
-          .then((list) => {
-            if (seq === seqRef.current) setItems(list);
-          })
-          .catch((err: unknown) => {
-            if (seq === seqRef.current)
-              setError(err instanceof Error ? err.message : String(err));
-          })
-          .finally(() => {
-            if (seq === seqRef.current) setLoading(false);
-          });
-      }, delay);
-      return () => window.clearTimeout(timer);
-    },
-    [fetchFn, delay],
-  );
-
-  return { items, loading, error, search };
-}
-
-export function IssueTypeField({
-  value,
-  onChange,
-}: {
-  value?: string;
-  onChange: (id: string) => void;
-}) {
-  const t = useT();
-  const jiraAccount = useSettingsStore((s) => s.accounts.jira);
-  const [open, setOpen] = useState(false);
-  const [items, setItems] = useState<JiraIssueType[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const projectKey = jiraAccount?.projectKey;
-
-  useEffect(() => {
-    setItems([]);
-    setError(null);
-  }, [projectKey]);
-
-  useEffect(() => {
-    if (!open || !jiraAccount || !projectKey) return;
-    if (items.length > 0) return;
-    let cancelled = false;
-    setLoading(true);
-    setError(null);
-    sendBg<JiraIssueType[]>({
-      type: "jira.listIssueTypes",
-      projectKey,
-    })
-      .then((list) => !cancelled && setItems(list))
-      .catch((err: unknown) =>
-        !cancelled && setError(err instanceof Error ? err.message : String(err)),
-      )
-      .finally(() => !cancelled && setLoading(false));
-    return () => {
-      cancelled = true;
-    };
-  }, [open, jiraAccount, projectKey, items.length]);
-
-  const defaultId = jiraAccount?.issueTypeId;
-  const defaultName = jiraAccount?.issueTypeName;
-  const effectiveValue = value ?? defaultId;
-  const selected = items.find((i) => i.id === effectiveValue);
-
-  useEffect(() => {
-    if (!value && defaultId) onChange(defaultId);
-  }, [value, defaultId, onChange]);
-
-  return (
-    <FieldCombobox
-      open={open}
-      onOpenChange={setOpen}
-      loading={loading}
-      error={error}
-      placeholder={t("field.issueType.select")}
-      searchPlaceholder={t("field.issueType.search")}
-      emptyMessage={t("field.issueType.empty")}
-      label={selected?.name ?? (effectiveValue ? defaultName : undefined)}
-    >
-      {items.map((it) => (
-        <CommandItem
-          key={it.id}
-          value={it.name}
-          onSelect={() => {
-            onChange(it.id);
-            setOpen(false);
-          }}
-        >
-          <Check
-            className={cn(
-              "mr-2 h-4 w-4",
-              effectiveValue === it.id ? "opacity-100" : "opacity-0",
-            )}
-          />
-          {it.iconUrl ? (
-            <img src={it.iconUrl} alt="" className="mr-2 h-4 w-4" />
-          ) : null}
-          <span className="min-w-0 flex-1 truncate">{it.name}</span>
-        </CommandItem>
-      ))}
-    </FieldCombobox>
-  );
-}
-
-export function PriorityField({
-  value,
-  fallbackLabel,
-  onChange,
-}: {
-  value?: string;
-  fallbackLabel?: string;
-  onChange: (id: string | undefined, name?: string) => void;
-}) {
-  const t = useT();
-  const jiraAccount = useSettingsStore((s) => s.accounts.jira);
-  const [open, setOpen] = useState(false);
-  const [items, setItems] = useState<JiraPriority[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    if (!open || !jiraAccount) return;
-    if (items.length > 0) return;
-    let cancelled = false;
-    setLoading(true);
-    setError(null);
-    sendBg<JiraPriority[]>({ type: "jira.listPriorities" })
-      .then((list) => !cancelled && setItems(list))
-      .catch((err: unknown) =>
-        !cancelled && setError(err instanceof Error ? err.message : String(err)),
-      )
-      .finally(() => !cancelled && setLoading(false));
-    return () => {
-      cancelled = true;
-    };
-  }, [open, jiraAccount, items.length]);
-
-  const selected = items.find((i) => i.id === value);
-
-  return (
-    <FieldCombobox
-      open={open}
-      onOpenChange={setOpen}
-      loading={loading}
-      error={error}
-      placeholder={t("field.priority.select")}
-      searchPlaceholder={t("field.priority.search")}
-      emptyMessage={t("field.priority.empty")}
-      label={selected?.name}
-      fallbackLabel={fallbackLabel}
-      clearable={!!value}
-      onClear={() => onChange(undefined)}
-      groupLabel={t("field.priority.label")}
-    >
-      {items.map((p) => (
-        <CommandItem
-          key={p.id}
-          value={p.name}
-          onSelect={() => {
-            onChange(p.id, p.name);
-            setOpen(false);
-          }}
-        >
-          <Check
-            className={cn(
-              "mr-2 h-4 w-4",
-              value === p.id ? "opacity-100" : "opacity-0",
-            )}
-          />
-          {p.iconUrl ? (
-            <img src={p.iconUrl} alt="" className="mr-2 h-4 w-4" />
-          ) : null}
-          <span className="min-w-0 flex-1 truncate">{p.name}</span>
-        </CommandItem>
-      ))}
-    </FieldCombobox>
-  );
-}
-
-export function AssigneeField({
-  value,
-  fallbackLabel,
-  onChange,
-}: {
-  value?: string;
-  fallbackLabel?: string;
-  onChange: (id: string | undefined, name?: string) => void;
-}) {
-  const t = useT();
-  const jira = useJiraConfig();
-  const [open, setOpen] = useState(false);
-
-  const fetchUsers = useCallback(
-    (query: string) => {
-      if (!jira) return Promise.resolve([]);
-      return sendBg<JiraUser[]>({
-        type: "jira.searchUsers",
-        query,
-      });
-    },
-    [jira],
-  );
-
-  const { items, loading, error, search } = useDebouncedSearch(fetchUsers);
-
-  useEffect(() => {
-    if (open) return search("");
-  }, [open, search]);
-
-  const selected = items.find((u) => u.accountId === value);
-
-  return (
-    <FieldCombobox
-      open={open}
-      onOpenChange={setOpen}
-      loading={loading}
-      error={error}
-      placeholder={t("field.assignee.select")}
-      searchPlaceholder={t("field.assignee.search")}
-      emptyMessage={t("field.assignee.empty")}
-      label={selected?.displayName}
-      fallbackLabel={fallbackLabel}
-      clearable={!!value}
-      onClear={() => onChange(undefined)}
-      onSearch={search}
-      groupLabel={t("field.assignee.label")}
-    >
-      {items.map((u) => (
-        <CommandItem
-          key={u.accountId}
-          value={u.displayName}
-          onSelect={() => {
-            onChange(u.accountId, u.displayName);
-            setOpen(false);
-          }}
-        >
-          <Check
-            className={cn(
-              "mr-2 h-4 w-4",
-              value === u.accountId ? "opacity-100" : "opacity-0",
-            )}
-          />
-          {u.avatarUrls?.["16x16"] ? (
-            <img
-              src={u.avatarUrls["16x16"]}
-              alt=""
-              className="mr-2 h-4 w-4 rounded-full"
-            />
-          ) : null}
-          <span className="min-w-0 flex-1 truncate">{u.displayName}</span>
-        </CommandItem>
-      ))}
-    </FieldCombobox>
-  );
-}
-
-export function EpicField({
-  value,
-  fallbackLabel,
-  onChange,
-  hierarchyLevels,
-}: {
-  value?: string;
-  fallbackLabel?: string;
-  onChange: (key: string | undefined, label?: string) => void;
-  hierarchyLevels?: number[];
-}) {
-  const t = useT();
-  const jira = useJiraConfig();
-  const [open, setOpen] = useState(false);
-
-  const fetchEpics = useCallback(
-    (query: string) => {
-      if (!jira) return Promise.resolve([]);
-      return sendBg<JiraIssueSummary[]>({
-        type: "jira.searchEpics",
-        projectKey: jira.projectKey,
-        query: query || undefined,
-        hierarchyLevels,
-      });
-    },
-    [jira, hierarchyLevels],
-  );
-
-  const { items, loading, error, search } = useDebouncedSearch(fetchEpics);
-
-  useEffect(() => {
-    if (open) return search("");
-  }, [open, search]);
-
-  const selected = items.find((i) => i.key === value);
-
-  return (
-    <FieldCombobox
-      open={open}
-      onOpenChange={setOpen}
-      loading={loading}
-      error={error}
-      placeholder={t("field.epic.select")}
-      searchPlaceholder={t("field.epic.search")}
-      emptyMessage={t("field.epic.empty")}
-      label={selected ? `${selected.key} ${selected.fields.summary}` : undefined}
-      fallbackLabel={fallbackLabel}
-      clearable={!!value}
-      onClear={() => onChange(undefined)}
-      onSearch={search}
-      groupLabel={t("field.epic.label")}
-    >
-      {items.map((epic) => (
-        <CommandItem
-          key={epic.id}
-          value={`${epic.key} ${epic.fields.summary}`}
-          onSelect={() => {
-            onChange(epic.key, `${epic.key} ${epic.fields.summary}`);
-            setOpen(false);
-          }}
-        >
-          <Check
-            className={cn(
-              "mr-2 h-4 w-4",
-              value === epic.key ? "opacity-100" : "opacity-0",
-            )}
-          />
-          <span className="shrink-0 text-muted-foreground">{epic.key}</span>
-          <span className="ml-1.5 min-w-0 flex-1 truncate">{epic.fields.summary}</span>
-        </CommandItem>
-      ))}
-    </FieldCombobox>
-  );
-}
-
-function FieldCombobox({
-  open,
-  onOpenChange,
-  loading,
-  error,
-  placeholder,
-  searchPlaceholder,
-  emptyMessage,
-  label,
-  fallbackLabel,
-  clearable,
-  onClear,
-  onSearch,
-  groupLabel,
-  children,
-}: {
-  open: boolean;
-  onOpenChange: (v: boolean) => void;
-  loading: boolean;
-  error: string | null;
-  placeholder: string;
-  searchPlaceholder: string;
-  emptyMessage: string;
-  label?: string;
-  fallbackLabel?: string;
-  clearable?: boolean;
-  onClear?: () => void;
-  onSearch?: (query: string) => void;
-  groupLabel?: string;
-  children: React.ReactNode;
-}) {
-  const t = useT();
-  return (
-    <Popover open={open} onOpenChange={onOpenChange}>
-      <PopoverTrigger asChild>
-        <Button
-          variant="outline"
-          role="combobox"
-          aria-expanded={open}
-          className="w-full justify-between font-normal"
-        >
-          <span
-            className={cn(
-              "min-w-0 flex-1 truncate text-left",
-              !label && !fallbackLabel && "text-muted-foreground",
-            )}
-          >
-            {label || fallbackLabel || placeholder}
-          </span>
-          <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-        </Button>
-      </PopoverTrigger>
-      <PopoverContent
-        className="w-[var(--radix-popover-trigger-width)] p-0"
-        onWheel={(e) => e.stopPropagation()}
-      >
-        <Command shouldFilter={!onSearch}>
-          <CommandInput
-            placeholder={searchPlaceholder}
-            onValueChange={onSearch}
-          />
-          <CommandList>
-            {loading ? (
-              <div className="flex items-center justify-center gap-2 py-6 text-xs text-muted-foreground">
-                <Loader2 className="h-3 w-3 animate-spin" />
-                {t("common.loading")}
-              </div>
-            ) : error ? (
-              <div className="px-3 py-6 text-center text-xs text-destructive">
-                {error}
-              </div>
-            ) : (
-              <>
-                {clearable && onClear ? (
-                  <CommandGroup heading={t("common.actions")}>
-                    <CommandItem
-                      value="__clear__"
-                      onSelect={() => {
-                        onClear();
-                        onOpenChange(false);
-                      }}
-                    >
-                      <X className="h-3.5 w-3.5" />
-                      <span className="text-xs">{t("common.deselect")}</span>
-                    </CommandItem>
-                  </CommandGroup>
-                ) : null}
-                <CommandEmpty>{emptyMessage}</CommandEmpty>
-                <CommandGroup heading={groupLabel}>{children}</CommandGroup>
-              </>
-            )}
-          </CommandList>
-        </Command>
-      </PopoverContent>
-    </Popover>
   );
 }
