+++ src/sidepanel/tabs/IssueListTab.tsx
@@ -1,13 +1,6 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { ArrowUpRight, Check, ChevronDown, CircleCheck, Inbox, Loader2, Search, SearchX, Trash2, X } from "lucide-react";
-import { toast } from "sonner";
-import {
-  SiGithub,
-  SiJirasoftware,
-  SiLinear,
-  SiNotion,
-} from "@icons-pack/react-simple-icons";
-import { useT, dateBcp47, type TranslationFn } from "@/i18n";
+import { useCallback, useMemo, useRef, useState } from "react";
+import { ArrowUpRight, CircleCheck, Inbox, Loader2, Search, SearchX, X } from "lucide-react";
+import { useT } from "@/i18n";
 import {
   AlertDialog,
   AlertDialogAction,
@@ -21,92 +14,21 @@ import {
 } from "@/components/ui/alert-dialog";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import { Card, CardContent } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
 import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { useIssuesStore, type IssueRecord } from "@/store/issues-store";
-import { useSettingsStore, jiraSiteId } from "@/store/settings-store";
-import type { JiraIssueStatus, JiraTransition } from "@/types/jira";
-import type { GithubIssueStatus } from "@/types/github";
-import type { LinearIssueStatus, LinearWorkflowState } from "@/types/linear";
-import type { NotionDatabaseSchema, NotionPageStatus } from "@/types/notion";
-import type { NormalizedSubmitResult, PlatformId } from "@/types/platform";
-import { BgError, sendBg } from "@/types/messages";
-import { extractNotionPageId } from "@/lib/notion-page-id";
+import { useIssuesStore } from "@/store/issues-store";
+import type { NormalizedSubmitResult } from "@/types/platform";
 import { PageFooter, PageScroll, PageShell, Section } from "@/sidepanel/components/Section";
 import { DraftDetailDialog } from "./DraftDetailDialog";
-import { notionStatusCategory } from "./notionStatusColors";
-
-type StatusFilter = "all" | "submitted" | "draft";
-
-export function isRefreshable(issue: IssueRecord): boolean {
-  if (issue.status !== "submitted" || !issue.url || !issue.key) return false;
-  if (issue.platform === "jira") return true;
-  if (issue.platform === "github") {
-    return !!resolveGithubCoords(issue);
-  }
-  if (issue.platform === "linear") return true;
-  if (issue.platform === "notion") {
-    return !!resolveNotionPageId(issue);
-  }
-  return false;
-}
-
-export function resolveNotionPageId(
-  issue: Pick<IssueRecord, "notionPageId" | "url">,
-): string | null {
-  if (issue.notionPageId) return issue.notionPageId;
-  return extractNotionPageId(issue.url);
-}
-
-export function parseGithubIssueNumber(key: string | undefined): number | null {
-  if (!key) return null;
-  const m = key.match(/^#?(\d+)$/);
-  return m ? Number(m[1]) : null;
-}
-
-// `https://github.com/{owner}/{repo}/issues/{number}` 형태에서 좌표 추출. 다른 패턴이면 null.
-export function parseGithubIssueUrl(
-  url: string | undefined,
-): { owner: string; repo: string; number: number } | null {
-  if (!url) return null;
-  try {
-    const u = new URL(url);
-    if (u.hostname !== "github.com") return null;
-    const m = u.pathname.match(/^\/([^/]+)\/([^/]+)\/issues\/(\d+)\/?$/);
-    if (!m) return null;
-    return { owner: m[1], repo: m[2], number: Number(m[3]) };
-  } catch {
-    return null;
-  }
-}
-
-// IssueRecord에 githubOwner/githubRepo가 없는 구 entry는 url에서 fallback 파싱.
-export function resolveGithubCoords(
-  issue: Pick<IssueRecord, "githubOwner" | "githubRepo" | "key" | "url">,
-): { owner: string; repo: string; number: number } | null {
-  const fromUrl = parseGithubIssueUrl(issue.url);
-  const number = parseGithubIssueNumber(issue.key) ?? fromUrl?.number ?? null;
-  const owner = issue.githubOwner ?? fromUrl?.owner ?? null;
-  const repo = issue.githubRepo ?? fromUrl?.repo ?? null;
-  if (!owner || !repo || number == null) return null;
-  return { owner, repo, number };
-}
-
-export function matchesQuery(issue: IssueRecord, q: string): boolean {
-  const lower = q.toLowerCase();
-  return (
-    issue.title.toLowerCase().includes(lower) ||
-    issue.pageUrl.toLowerCase().includes(lower) ||
-    (issue.key?.toLowerCase().includes(lower) ?? false)
-  );
-}
-
-export function matchesStatus(issue: IssueRecord, filter: StatusFilter): boolean {
-  if (filter === "all") return true;
-  return issue.status === filter;
-}
+import { IssueRow } from "./IssueRow";
+import {
+  groupByDate,
+  isRefreshable,
+  issueTimestamp,
+  matchesQuery,
+  matchesStatus,
+  type StatusFilter,
+} from "./issueListUtils";
 
 export function IssueListTab() {
   const t = useT();
@@ -318,988 +240,4 @@ export function IssueListTab() {
       />
     </PageShell>
   );
-}
-
-function IssueRow({
-  issue,
-  refreshKey,
-  onOpenDraft,
-  onBadgeLoaded,
-}: {
-  issue: IssueRecord;
-  refreshKey: number;
-  onOpenDraft: () => void;
-  onBadgeLoaded: () => void;
-}) {
-  const t = useT();
-  const isSubmitted = issue.status === "submitted" && !!issue.url;
-  const removeIssue = useIssuesStore((s) => s.removeIssue);
-  const [badgeHover, setBadgeHover] = useState(false);
-
-  const textMetaParts: string[] = [];
-  if (isSubmitted) {
-    textMetaParts.push(formatDate(issueTimestamp(issue), t));
-    if (issue.key) textMetaParts.push(formatIssueKey(issue));
-  } else {
-    textMetaParts.push(t("issueList.draft"));
-    textMetaParts.push(formatDate(issueTimestamp(issue), t));
-  }
-
-  const handleCardClick = () => {
-    if (isSubmitted) {
-      chrome.tabs.create({ url: issue.url!, active: true });
-    } else {
-      // Card는 비포커서블이라 직전 포커스(탭 trigger 등)가 그대로 남는다.
-      // Radix Dialog가 root에 aria-hidden을 씌우면 포커스된 후손이 있어
-      // 브라우저 a11y 경고가 뜨므로, 미리 blur해서 경고를 막는다.
-      if (document.activeElement instanceof HTMLElement && document.activeElement !== document.body) {
-        document.activeElement.blur();
-      }
-      onOpenDraft();
-    }
-  };
-
-  return (
-    <li>
-      <Card
-        className={`group cursor-pointer transition-colors ${badgeHover ? "" : "hover:bg-muted/50"}`}
-        onClick={handleCardClick}
-      >
-        <CardContent className="flex items-center justify-between gap-3 px-4 py-3">
-          <div className="flex min-w-0 flex-col">
-            <span className="truncate text-base font-medium text-foreground">
-              {issue.title || t("common.untitled")}
-            </span>
-            <span className="flex min-w-0 items-center gap-1.5 text-sm text-muted-foreground">
-              {isSubmitted ? (
-                <>
-                  <PlatformChip platform={issue.platform} />
-                  {textMetaParts.length > 0 ? <span aria-hidden>·</span> : null}
-                </>
-              ) : null}
-              <span className="min-w-0 truncate">{textMetaParts.join(" · ")}</span>
-            </span>
-          </div>
-          {isSubmitted && issue.key ? (
-            <span
-              onClick={(e) => e.stopPropagation()}
-              onMouseEnter={() => setBadgeHover(true)}
-              onMouseLeave={() => setBadgeHover(false)}
-            >
-              <SubmittedBadge
-                issueId={issue.id}
-                issueKey={issue.key}
-                issueSiteId={issue.jiraSiteId}
-                issueUrl={issue.url}
-                platform={issue.platform}
-                githubOwner={issue.githubOwner}
-                githubRepo={issue.githubRepo}
-                linearIdentifier={issue.linearIdentifier}
-                notionPageId={issue.notionPageId}
-                notionDatabaseId={issue.notionDatabaseId}
-                refreshKey={refreshKey}
-                onLoaded={onBadgeLoaded}
-              />
-            </span>
-          ) : (
-            <AlertDialog>
-              <AlertDialogTrigger asChild>
-                <Button
-                  variant="outline"
-                  size="icon"
-                  className="h-8 w-8 shrink-0 text-muted-foreground hover:text-destructive"
-                  onClick={(e) => e.stopPropagation()}
-                >
-                  <Trash2 />
-                </Button>
-              </AlertDialogTrigger>
-              <AlertDialogContent onClick={(e) => e.stopPropagation()}>
-                <AlertDialogHeader>
-                  <AlertDialogTitle>{t("issueList.deleteDraft.title")}</AlertDialogTitle>
-                  <AlertDialogDescription>
-                    {t("issueList.deleteDraft.body")}
-                  </AlertDialogDescription>
-                </AlertDialogHeader>
-                <AlertDialogFooter>
-                  <AlertDialogCancel>{t("common.close")}</AlertDialogCancel>
-                  <AlertDialogAction
-                    onClick={() => removeIssue(issue.id)}
-                  >
-                    {t("issueList.deleteIssue")}
-                  </AlertDialogAction>
-                </AlertDialogFooter>
-              </AlertDialogContent>
-            </AlertDialog>
-          )}
-        </CardContent>
-      </Card>
-    </li>
-  );
-}
-
-const STATUS_CATEGORY_COLORS: Record<
-  string,
-  { bg: string; text: string; darkBg: string; darkText: string }
-> = {
-  new: {
-    bg: "bg-slate-100",
-    text: "text-slate-700",
-    darkBg: "dark:bg-slate-500/15",
-    darkText: "dark:text-slate-300",
-  },
-  indeterminate: {
-    bg: "bg-blue-100",
-    text: "text-blue-700",
-    darkBg: "dark:bg-blue-500/15",
-    darkText: "dark:text-blue-300",
-  },
-  done: {
-    bg: "bg-green-100",
-    text: "text-green-700",
-    darkBg: "dark:bg-green-500/15",
-    darkText: "dark:text-green-300",
-  },
-};
-
-const LINEAR_STATE_TYPE_COLORS: Record<string, typeof STATUS_CATEGORY_COLORS[string]> = {
-  backlog: STATUS_CATEGORY_COLORS.new,
-  unstarted: STATUS_CATEGORY_COLORS.new,
-  started: STATUS_CATEGORY_COLORS.indeterminate,
-  completed: STATUS_CATEGORY_COLORS.done,
-  cancelled: STATUS_CATEGORY_COLORS.new,
-};
-
-const LINEAR_STATE_I18N: Record<string, string> = {
-  backlog: "issueList.linear.backlog",
-  unstarted: "issueList.linear.unstarted",
-  started: "issueList.linear.started",
-  completed: "issueList.linear.completed",
-  cancelled: "issueList.linear.cancelled",
-};
-
-
-function PlatformChip({ platform }: { platform: PlatformId }) {
-  const t = useT();
-  if (platform === "jira") {
-    return (
-      <span className="inline-flex shrink-0 items-center gap-1">
-        <SiJirasoftware className="h-3 w-3" color="default" />
-        {t("platform.tab.jira")}
-      </span>
-    );
-  }
-  if (platform === "linear") {
-    return (
-      <span className="inline-flex shrink-0 items-center gap-1">
-        <SiLinear className="h-3 w-3" color="default" />
-        {t("platform.tab.linear")}
-      </span>
-    );
-  }
-  if (platform === "notion") {
-    return (
-      <span className="inline-flex shrink-0 items-center gap-1">
-        <SiNotion className="h-3 w-3 dark:invert" color="default" />
-        {t("platform.tab.notion")}
-      </span>
-    );
-  }
-  return (
-    <span className="inline-flex shrink-0 items-center gap-1">
-      <SiGithub className="h-3 w-3 dark:invert" color="default" />
-      {t("platform.tab.github")}
-    </span>
-  );
-}
-
-type GithubBadgeStatus =
-  | { kind: "open" }
-  | { kind: "closed"; reason: "completed" | "not_planned" | "reopened" | null };
-
-type GithubTargetState = "open" | "closed_completed" | "closed_not_planned";
-
-function toGithubTargetState(s: GithubBadgeStatus): GithubTargetState {
-  if (s.kind === "open") return "open";
-  return s.reason === "not_planned" ? "closed_not_planned" : "closed_completed";
-}
-
-function GithubStatusBadge({
-  ghStatus,
-  issueId,
-  owner,
-  repo,
-  number,
-  onStatusChanged,
-}: {
-  ghStatus: GithubBadgeStatus;
-  issueId: string;
-  owner: string;
-  repo: string;
-  number: number;
-  onStatusChanged: (s: GithubBadgeStatus) => void;
-}) {
-  const t = useT();
-  const patchIssue = useIssuesStore((s) => s.patchIssue);
-  const [open, setOpen] = useState(false);
-  const [updating, setUpdating] = useState(false);
-
-  const current = toGithubTargetState(ghStatus);
-
-  const options: { key: GithubTargetState; label: string; colors: typeof STATUS_CATEGORY_COLORS[string] }[] = [
-    { key: "open", label: t("issueList.github.status.open"), colors: STATUS_CATEGORY_COLORS.indeterminate },
-    { key: "closed_completed", label: t("issueList.github.status.closedCompleted"), colors: STATUS_CATEGORY_COLORS.done },
-    { key: "closed_not_planned", label: t("issueList.github.status.closedNotPlanned"), colors: STATUS_CATEGORY_COLORS.new },
-  ];
-
-  const currentOption = options.find((o) => o.key === current)!;
-
-  const handleSelect = (target: GithubTargetState) => {
-    if (target === current) { setOpen(false); return; }
-    setOpen(false);
-    setUpdating(true);
-    const state = target === "open" ? "open" as const : "closed" as const;
-    const stateReason = target === "closed_not_planned" ? "not_planned" as const : target === "closed_completed" ? "completed" as const : null;
-    sendBg<GithubIssueStatus>({
-      type: "github.updateIssueState",
-      owner, repo, number, state,
-      ...(stateReason ? { stateReason } : {}),
-    })
-      .then((res) => {
-        const newStatus: GithubBadgeStatus =
-          res.state === "open"
-            ? { kind: "open" }
-            : { kind: "closed", reason: res.stateReason ?? null };
-        onStatusChanged(newStatus);
-        const patch: Partial<IssueRecord> = {};
-        if (res.title) patch.title = res.title;
-        patch.githubLabels = res.labels.map((l) => l.name).filter(Boolean);
-        if (Object.keys(patch).length) patchIssue(issueId, patch);
-      })
-      .catch(() => toast.error(t("issueList.github.statusUpdateFailed")))
-      .finally(() => setUpdating(false));
-  };
-
-  return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <PopoverTrigger asChild>
-        <button
-          type="button"
-          className="shrink-0 rounded-md focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-          onClick={(e) => e.stopPropagation()}
-          disabled={updating}
-        >
-          <Badge
-            variant="outline"
-            className={`relative w-fit border-transparent text-[11px] ${currentOption.colors.bg} ${currentOption.colors.text} ${currentOption.colors.darkBg} ${currentOption.colors.darkText} ${updating ? "opacity-50" : ""}`}
-          >
-            {updating && <Loader2 className="mr-1 h-3 w-3 animate-spin" />}
-            {currentOption.label}
-            {!updating && <ChevronDown className="ml-0.5 !size-3.5" />}
-          </Badge>
-        </button>
-      </PopoverTrigger>
-      <PopoverContent
-        className="w-auto p-1"
-        align="end"
-        onClick={(e) => e.stopPropagation()}
-      >
-        {options.map((opt) => (
-          <button
-            key={opt.key}
-            type="button"
-            className="flex w-full items-center gap-2 rounded-sm px-2 py-1.5 hover:bg-accent hover:text-accent-foreground"
-            onClick={() => handleSelect(opt.key)}
-          >
-            <Check className={`h-3.5 w-3.5 shrink-0 ${opt.key === current ? "opacity-100" : "opacity-0"}`} />
-            <Badge
-              variant="outline"
-              className={`border-transparent text-[11px] ${opt.colors.bg} ${opt.colors.text} ${opt.colors.darkBg} ${opt.colors.darkText}`}
-            >
-              {opt.label}
-            </Badge>
-          </button>
-        ))}
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function JiraStatusBadge({
-  issueKey,
-  issueId,
-  currentStatus,
-  onStatusChanged,
-}: {
-  issueKey: string;
-  issueId: string;
-  currentStatus: JiraIssueStatus;
-  onStatusChanged: (s: JiraIssueStatus) => void;
-}) {
-  const t = useT();
-  const patchIssue = useIssuesStore((s) => s.patchIssue);
-  const [open, setOpen] = useState(false);
-  const [transitions, setTransitions] = useState<JiraTransition[] | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [updating, setUpdating] = useState(false);
-
-  const colors = STATUS_CATEGORY_COLORS[currentStatus.categoryKey] ?? STATUS_CATEGORY_COLORS.new;
-
-  const handleOpenChange = (v: boolean) => {
-    setOpen(v);
-    if (v && !transitions) {
-      setLoading(true);
-      sendBg<JiraTransition[]>({ type: "jira.getTransitions", issueKey })
-        .then(setTransitions)
-        .catch(() => setTransitions([]))
-        .finally(() => setLoading(false));
-    }
-    if (!v) setTransitions(null);
-  };
-
-  const handleSelect = (tr: JiraTransition) => {
-    if (updating) return;
-    setOpen(false);
-    setUpdating(true);
-    sendBg<JiraIssueStatus>({
-      type: "jira.transitionIssue",
-      issueKey,
-      transitionId: tr.id,
-    })
-      .then((res) => {
-        onStatusChanged(res);
-        const patch: Partial<IssueRecord> = {};
-        if (res.issueTypeName) patch.issueTypeName = res.issueTypeName;
-        if (res.summary) patch.title = res.summary;
-        if (Object.keys(patch).length) patchIssue(issueId, patch);
-      })
-      .catch((err) => {
-        if (err instanceof BgError && err.status === 400) {
-          toast.error(t("issueList.jira.requiredFieldsError"));
-        } else {
-          toast.error(t("issueList.jira.statusUpdateFailed"));
-        }
-      })
-      .finally(() => setUpdating(false));
-  };
-
-  return (
-    <Popover open={open} onOpenChange={handleOpenChange}>
-      <PopoverTrigger asChild>
-        <button
-          type="button"
-          className="shrink-0 rounded-md focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-          onClick={(e) => e.stopPropagation()}
-          disabled={updating}
-        >
-          <Badge
-            variant="outline"
-            className={`relative w-fit border-transparent text-[11px] ${colors.bg} ${colors.text} ${colors.darkBg} ${colors.darkText} ${updating ? "opacity-50" : ""}`}
-          >
-            {updating && <Loader2 className="mr-1 h-3 w-3 animate-spin" />}
-            {currentStatus.name}
-            {!updating && <ChevronDown className="ml-0.5 !size-3.5" />}
-          </Badge>
-        </button>
-      </PopoverTrigger>
-      <PopoverContent
-        className="w-auto max-h-[300px] overflow-y-auto p-1"
-        align="end"
-        onClick={(e) => e.stopPropagation()}
-      >
-        {loading ? (
-          <div className="flex items-center justify-center px-4 py-2">
-            <Loader2 className="h-4 w-4 animate-spin text-muted-foreground" />
-          </div>
-        ) : transitions && transitions.length === 0 ? (
-          <div className="px-3 py-2 text-sm text-muted-foreground">
-            {t("issueList.jira.noTransitions")}
-          </div>
-        ) : (
-          transitions?.map((tr) => {
-            const trColors = STATUS_CATEGORY_COLORS[tr.to.categoryKey] ?? STATUS_CATEGORY_COLORS.new;
-            return (
-              <button
-                key={tr.id}
-                type="button"
-                className="flex w-full items-center gap-2 rounded-sm px-2 py-1.5 hover:bg-accent hover:text-accent-foreground"
-                onClick={() => handleSelect(tr)}
-              >
-                <Check className={`h-3.5 w-3.5 shrink-0 ${tr.to.name === currentStatus.name ? "opacity-100" : "opacity-0"}`} />
-                <Badge
-                  variant="outline"
-                  className={`border-transparent text-[11px] ${trColors.bg} ${trColors.text} ${trColors.darkBg} ${trColors.darkText}`}
-                >
-                  {tr.to.name}
-                </Badge>
-              </button>
-            );
-          })
-        )}
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function LinearStatusBadge({
-  issueId,
-  issueIdentifier,
-  currentState,
-  onStatusChanged,
-}: {
-  issueId: string;
-  issueIdentifier: string;
-  currentState: { name: string; type: string };
-  onStatusChanged: (s: LinearIssueStatus) => void;
-}) {
-  const t = useT();
-  const patchIssue = useIssuesStore((s) => s.patchIssue);
-  const [open, setOpen] = useState(false);
-  const [states, setStates] = useState<LinearWorkflowState[] | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [updating, setUpdating] = useState(false);
-
-  const stateColors = LINEAR_STATE_TYPE_COLORS[currentState.type] ?? STATUS_CATEGORY_COLORS.new;
-  const i18nKey = LINEAR_STATE_I18N[currentState.type] as Parameters<typeof t>[0] | undefined;
-  const currentLabel = i18nKey ? t(i18nKey) : currentState.name;
-
-  const handleOpenChange = (v: boolean) => {
-    setOpen(v);
-    if (v && !states) {
-      setLoading(true);
-      sendBg<LinearWorkflowState[]>({ type: "linear.getWorkflowStates", issueIdentifier })
-        .then(setStates)
-        .catch(() => setStates([]))
-        .finally(() => setLoading(false));
-    }
-    if (!v) setStates(null);
-  };
-
-  const handleSelect = (state: LinearWorkflowState) => {
-    if (updating || state.name === currentState.name) return;
-    setOpen(false);
-    setUpdating(true);
-    sendBg<LinearIssueStatus>({
-      type: "linear.updateIssueState",
-      issueId,
-      stateId: state.id,
-    })
-      .then((res) => {
-        onStatusChanged(res);
-        const patch: Partial<IssueRecord> = {};
-        if (res.title) patch.title = res.title;
-        if (res.identifier) patch.linearIdentifier = res.identifier;
-        if (res.labels.length > 0) patch.linearLabelName = res.labels[0].name;
-        if (Object.keys(patch).length) patchIssue(issueId, patch);
-      })
-      .catch(() => toast.error(t("issueList.linear.statusUpdateFailed")))
-      .finally(() => setUpdating(false));
-  };
-
-  return (
-    <Popover open={open} onOpenChange={handleOpenChange}>
-      <PopoverTrigger asChild>
-        <button
-          type="button"
-          className="shrink-0 rounded-md focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-          onClick={(e) => e.stopPropagation()}
-          disabled={updating}
-        >
-          <Badge
-            variant="outline"
-            className={`relative w-fit border-transparent text-[11px] ${stateColors.bg} ${stateColors.text} ${stateColors.darkBg} ${stateColors.darkText} ${updating ? "opacity-50" : ""}`}
-          >
-            {updating && <Loader2 className="mr-1 h-3 w-3 animate-spin" />}
-            {currentLabel}
-            {!updating && <ChevronDown className="ml-0.5 !size-3.5" />}
-          </Badge>
-        </button>
-      </PopoverTrigger>
-      <PopoverContent
-        className="w-auto max-h-[300px] overflow-y-auto p-1"
-        align="end"
-        onClick={(e) => e.stopPropagation()}
-      >
-        {loading ? (
-          <div className="flex items-center justify-center px-4 py-2">
-            <Loader2 className="h-4 w-4 animate-spin text-muted-foreground" />
-          </div>
-        ) : (
-          states?.map((st) => {
-            const stColors = LINEAR_STATE_TYPE_COLORS[st.type] ?? STATUS_CATEGORY_COLORS.new;
-            return (
-              <button
-                key={st.id}
-                type="button"
-                className="flex w-full items-center gap-2 rounded-sm px-2 py-1.5 hover:bg-accent hover:text-accent-foreground"
-                onClick={() => handleSelect(st)}
-              >
-                <Check className={`h-3.5 w-3.5 shrink-0 ${st.name === currentState.name ? "opacity-100" : "opacity-0"}`} />
-                <Badge
-                  variant="outline"
-                  className={`border-transparent text-[11px] ${stColors.bg} ${stColors.text} ${stColors.darkBg} ${stColors.darkText}`}
-                >
-                  {st.name}
-                </Badge>
-              </button>
-            );
-          })
-        )}
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function NotionStatusBadge({
-  pageId,
-  databaseId,
-  issueId,
-  currentOption,
-  onStatusChanged,
-}: {
-  pageId: string;
-  databaseId: string;
-  issueId: string;
-  currentOption: { name: string; color: string };
-  onStatusChanged: (s: NotionPageStatus) => void;
-}) {
-  const t = useT();
-  const patchIssue = useIssuesStore((s) => s.patchIssue);
-  const [open, setOpen] = useState(false);
-  const [options, setOptions] = useState<Array<{ id: string; name: string; color: string }> | null>(null);
-  const [statusPropertyName, setStatusPropertyName] = useState<string | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [updating, setUpdating] = useState(false);
-
-  const category = notionStatusCategory(currentOption.color);
-  const colors = STATUS_CATEGORY_COLORS[category];
-
-  const handleOpenChange = (v: boolean) => {
-    setOpen(v);
-    if (v && !options) {
-      setLoading(true);
-      sendBg<NotionDatabaseSchema>({ type: "notion.getDatabaseSchema", databaseId })
-        .then((schema) => {
-          setOptions(schema.statusProperty?.options ?? []);
-          setStatusPropertyName(schema.statusProperty?.name ?? null);
-        })
-        .catch(() => setOptions([]))
-        .finally(() => setLoading(false));
-    }
-    if (!v) { setOptions(null); setStatusPropertyName(null); }
-  };
-
-  const handleSelect = (optionName: string) => {
-    if (updating || optionName === currentOption.name || !statusPropertyName) return;
-    setOpen(false);
-    setUpdating(true);
-    sendBg<NotionPageStatus>({
-      type: "notion.updatePageStatus",
-      pageId,
-      propertyName: statusPropertyName,
-      optionName,
-    })
-      .then((res) => {
-        onStatusChanged(res);
-        const patch: Partial<IssueRecord> = {};
-        if (res.title) patch.title = res.title;
-        if (res.statusOption) patch.notionStatusOption = res.statusOption.name;
-        if (Object.keys(patch).length) patchIssue(issueId, patch);
-      })
-      .catch(() => toast.error(t("issueList.notion.statusUpdateFailed")))
-      .finally(() => setUpdating(false));
-  };
-
-  return (
-    <Popover open={open} onOpenChange={handleOpenChange}>
-      <PopoverTrigger asChild>
-        <button
-          type="button"
-          className="shrink-0 rounded-md focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-          onClick={(e) => e.stopPropagation()}
-          disabled={updating}
-        >
-          <Badge
-            variant="outline"
-            className={`relative w-fit border-transparent text-[11px] ${colors.bg} ${colors.text} ${colors.darkBg} ${colors.darkText} ${updating ? "opacity-50" : ""}`}
-          >
-            {updating && <Loader2 className="mr-1 h-3 w-3 animate-spin" />}
-            {currentOption.name}
-            {!updating && <ChevronDown className="ml-0.5 !size-3.5" />}
-          </Badge>
-        </button>
-      </PopoverTrigger>
-      <PopoverContent
-        className="w-auto max-h-[300px] overflow-y-auto p-1"
-        align="end"
-        onClick={(e) => e.stopPropagation()}
-      >
-        {loading ? (
-          <div className="flex items-center justify-center px-4 py-2">
-            <Loader2 className="h-4 w-4 animate-spin text-muted-foreground" />
-          </div>
-        ) : options && options.length === 0 ? (
-          <div className="px-3 py-2 text-sm text-muted-foreground">
-            {t("issueList.notion.noStatusOptions")}
-          </div>
-        ) : (
-          options?.map((opt) => {
-            const optCategory = notionStatusCategory(opt.color);
-            const optColors = STATUS_CATEGORY_COLORS[optCategory];
-            return (
-              <button
-                key={opt.id}
-                type="button"
-                className="flex w-full items-center gap-2 rounded-sm px-2 py-1.5 hover:bg-accent hover:text-accent-foreground"
-                onClick={() => handleSelect(opt.name)}
-              >
-                <Check className={`h-3.5 w-3.5 shrink-0 ${opt.name === currentOption.name ? "opacity-100" : "opacity-0"}`} />
-                <Badge
-                  variant="outline"
-                  className={`border-transparent text-[11px] ${optColors.bg} ${optColors.text} ${optColors.darkBg} ${optColors.darkText}`}
-                >
-                  {opt.name}
-                </Badge>
-              </button>
-            );
-          })
-        )}
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function SubmittedBadge({
-  issueId,
-  issueKey,
-  issueSiteId,
-  issueUrl,
-  platform,
-  githubOwner,
-  githubRepo,
-  linearIdentifier,
-  notionPageId,
-  notionDatabaseId,
-  refreshKey,
-  onLoaded,
-}: {
-  issueId: string;
-  issueKey: string;
-  issueSiteId?: string;
-  issueUrl?: string;
-  platform: PlatformId;
-  githubOwner?: string;
-  githubRepo?: string;
-  linearIdentifier?: string;
-  notionPageId?: string;
-  notionDatabaseId?: string;
-  refreshKey: number;
-  onLoaded: () => void;
-}) {
-  const t = useT();
-  const jiraAccount = useSettingsStore((s) => s.accounts.jira);
-  const ghAccount = useSettingsStore((s) => s.accounts.github);
-  const linearAccount = useSettingsStore((s) => s.accounts.linear);
-  const notionAccount = useSettingsStore((s) => s.accounts.notion);
-  const patchIssue = useIssuesStore((s) => s.patchIssue);
-  const currentSiteId = jiraAccount?.auth ? jiraSiteId(jiraAccount.auth) : null;
-  const siteMatch = !issueSiteId || currentSiteId === issueSiteId;
-  const [jiraStatus, setJiraStatus] = useState<JiraIssueStatus | "error" | null>(null);
-  const [ghStatus, setGhStatus] = useState<GithubBadgeStatus | "error" | null>(null);
-  const [linearStatus, setLinearStatus] = useState<LinearIssueStatus | "error" | null>(null);
-  const [notionStatus, setNotionStatus] = useState<NotionPageStatus | "error" | null>(null);
-
-  const ghCoords = useMemo(
-    () => resolveGithubCoords({ githubOwner, githubRepo, key: issueKey, url: issueUrl }),
-    [githubOwner, githubRepo, issueKey, issueUrl],
-  );
-
-  useEffect(() => {
-    if (platform === "jira") {
-      if (!jiraAccount?.auth || !siteMatch) { setJiraStatus("error"); onLoaded(); return; }
-      sendBg<JiraIssueStatus>({ type: "jira.getIssueStatus", issueKey })
-        .then((res) => {
-          setJiraStatus(res);
-          const patch: Record<string, string> = {};
-          if (res.issueTypeName) patch.issueTypeName = res.issueTypeName;
-          if (res.summary) patch.title = res.summary;
-          if (Object.keys(patch).length) patchIssue(issueId, patch);
-        })
-        .catch(() => setJiraStatus("error"))
-        .finally(onLoaded);
-      return;
-    }
-    if (platform === "github") {
-      if (!ghAccount || !ghCoords) {
-        setGhStatus("error"); onLoaded(); return;
-      }
-      sendBg<GithubIssueStatus>({
-        type: "github.getIssueStatus",
-        owner: ghCoords.owner,
-        repo: ghCoords.repo,
-        number: ghCoords.number,
-      })
-        .then((res) => {
-          setGhStatus(
-            res.state === "open"
-              ? { kind: "open" }
-              : { kind: "closed", reason: res.stateReason ?? null },
-          );
-          const patch: Partial<IssueRecord> = {};
-          if (res.title) patch.title = res.title;
-          if (!githubOwner) patch.githubOwner = ghCoords.owner;
-          if (!githubRepo) patch.githubRepo = ghCoords.repo;
-          patch.githubLabels = res.labels.map((l) => l.name).filter(Boolean);
-          if (Object.keys(patch).length) patchIssue(issueId, patch);
-        })
-        .catch(() => setGhStatus("error"))
-        .finally(onLoaded);
-      return;
-    }
-    if (platform === "linear") {
-      const identifier = linearIdentifier ?? issueKey;
-      if (!linearAccount?.auth || !identifier) {
-        setLinearStatus("error"); onLoaded(); return;
-      }
-      sendBg<LinearIssueStatus>({ type: "linear.getIssueStatus", issueId: identifier })
-        .then((res) => {
-          setLinearStatus(res);
-          const patch: Partial<IssueRecord> = {};
-          if (res.title) patch.title = res.title;
-          if (res.identifier) patch.linearIdentifier = res.identifier;
-          if (res.labels.length > 0) patch.linearLabelName = res.labels[0].name;
-          if (Object.keys(patch).length) patchIssue(issueId, patch);
-        })
-        .catch(() => setLinearStatus("error"))
-        .finally(onLoaded);
-      return;
-    }
-    if (platform === "notion") {
-      const pageId =
-        resolveNotionPageId({ notionPageId, url: issueUrl }) ?? null;
-      if (!notionAccount?.auth || !pageId) {
-        setNotionStatus("error"); onLoaded(); return;
-      }
-      sendBg<NotionPageStatus>({ type: "notion.getPageStatus", pageId })
-        .then((res) => {
-          setNotionStatus(res);
-          const patch: Partial<IssueRecord> = {};
-          if (res.title) patch.title = res.title;
-          if (!notionPageId) patch.notionPageId = res.pageId;
-          if (res.statusOption) patch.notionStatusOption = res.statusOption.name;
-          if (Object.keys(patch).length) patchIssue(issueId, patch);
-        })
-        .catch(() => setNotionStatus("error"))
-        .finally(onLoaded);
-      return;
-    }
-    onLoaded();
-  }, [platform, jiraAccount?.auth, ghAccount, linearAccount?.auth, notionAccount?.auth, issueKey, issueUrl, ghCoords, linearIdentifier, notionPageId, refreshKey, siteMatch, onLoaded, issueId, patchIssue]);
-
-  if (platform === "jira") {
-    if (jiraStatus === "error") {
-      return (
-        <Badge variant="outline" className="w-fit shrink-0 text-[11px]">
-          {t("issueList.unknown")}
-        </Badge>
-      );
-    }
-    if (!jiraStatus) return null;
-    if (jiraAccount?.auth && siteMatch) {
-      return (
-        <JiraStatusBadge
-          issueKey={issueKey}
-          issueId={issueId}
-          currentStatus={jiraStatus}
-          onStatusChanged={setJiraStatus}
-        />
-      );
-    }
-    const colors = STATUS_CATEGORY_COLORS[jiraStatus.categoryKey] ?? STATUS_CATEGORY_COLORS.new;
-    return (
-      <Badge
-        variant="outline"
-        className={`w-fit shrink-0 border-transparent text-[11px] ${colors.bg} ${colors.text} ${colors.darkBg} ${colors.darkText}`}
-      >
-        {jiraStatus.name}
-      </Badge>
-    );
-  }
-
-  if (platform === "github") {
-    if (ghStatus === "error") {
-      return (
-        <Badge variant="outline" className="w-fit shrink-0 text-[11px]">
-          {t("issueList.unknown")}
-        </Badge>
-      );
-    }
-    if (!ghStatus) {
-      return (
-        <Badge variant="outline" className="w-fit shrink-0 text-[11px]">
-          {t("issueList.submitted")}
-        </Badge>
-      );
-    }
-    if (ghCoords) {
-      return (
-        <GithubStatusBadge
-          ghStatus={ghStatus}
-          issueId={issueId}
-          owner={ghCoords.owner}
-          repo={ghCoords.repo}
-          number={ghCoords.number}
-          onStatusChanged={setGhStatus}
-        />
-      );
-    }
-    const ghLabel =
-      ghStatus.kind === "open"
-        ? t("issueList.github.open")
-        : ghStatus.reason === "not_planned"
-          ? t("issueList.github.notPlanned")
-          : t("issueList.github.closed");
-    const ghColors =
-      ghStatus.kind === "open"
-        ? STATUS_CATEGORY_COLORS.indeterminate
-        : ghStatus.reason === "not_planned"
-          ? STATUS_CATEGORY_COLORS.new
-          : STATUS_CATEGORY_COLORS.done;
-    return (
-      <Badge
-        variant="outline"
-        className={`w-fit shrink-0 border-transparent text-[11px] ${ghColors.bg} ${ghColors.text} ${ghColors.darkBg} ${ghColors.darkText}`}
-      >
-        {ghLabel}
-      </Badge>
-    );
-  }
-
-  if (platform === "linear") {
-    if (linearStatus === "error") {
-      return (
-        <Badge variant="outline" className="w-fit shrink-0 text-[11px]">
-          {t("issueList.unknown")}
-        </Badge>
-      );
-    }
-    if (!linearStatus) {
-      return (
-        <Badge variant="outline" className="w-fit shrink-0 text-[11px]">
-          {t("issueList.submitted")}
-        </Badge>
-      );
-    }
-    if (linearAccount?.auth && linearStatus.id) {
-      return (
-        <LinearStatusBadge
-          issueId={linearStatus.id}
-          issueIdentifier={linearStatus.identifier}
-          currentState={linearStatus.state}
-          onStatusChanged={setLinearStatus}
-        />
-      );
-    }
-    const stateType = linearStatus.state.type;
-    const linearColors = LINEAR_STATE_TYPE_COLORS[stateType] ?? STATUS_CATEGORY_COLORS.new;
-    const i18nKey = LINEAR_STATE_I18N[stateType] as Parameters<typeof t>[0] | undefined;
-    const linearLabel = i18nKey ? t(i18nKey) : linearStatus.state.name;
-    return (
-      <Badge
-        variant="outline"
-        className={`w-fit shrink-0 border-transparent text-[11px] ${linearColors.bg} ${linearColors.text} ${linearColors.darkBg} ${linearColors.darkText}`}
-      >
-        {linearLabel}
-      </Badge>
-    );
-  }
-
-  // Notion
-  if (platform === "notion") {
-    if (notionStatus === "error") {
-      return (
-        <Badge variant="outline" className="w-fit shrink-0 text-[11px]">
-          {t("issueList.unknown")}
-        </Badge>
-      );
-    }
-    if (!notionStatus) {
-      return (
-        <Badge variant="outline" className="w-fit shrink-0 text-[11px]">
-          {t("issueList.submitted")}
-        </Badge>
-      );
-    }
-    if (notionStatus.statusOption) {
-      const resolvedPageId = resolveNotionPageId({ notionPageId, url: issueUrl }) ?? null;
-      if (notionAccount?.auth && notionDatabaseId && resolvedPageId) {
-        return (
-          <NotionStatusBadge
-            pageId={resolvedPageId}
-            databaseId={notionDatabaseId}
-            issueId={issueId}
-            currentOption={notionStatus.statusOption}
-            onStatusChanged={setNotionStatus}
-          />
-        );
-      }
-      const category = notionStatusCategory(notionStatus.statusOption.color);
-      const notionColors = STATUS_CATEGORY_COLORS[category];
-      return (
-        <Badge
-          variant="outline"
-          className={`w-fit shrink-0 border-transparent text-[11px] ${notionColors.bg} ${notionColors.text} ${notionColors.darkBg} ${notionColors.darkText}`}
-        >
-          {notionStatus.statusOption.name}
-        </Badge>
-      );
-    }
-    return (
-      <Badge variant="outline" className="w-fit shrink-0 text-[11px]">
-        {t("issueList.notion.noStatus")}
-      </Badge>
-    );
-  }
-
-  return null;
-}
-
-// Jira는 `[BUG-1]`, GitHub은 `#42`로 시각적 구분.
-// (GitHub key는 이미 `#`이 포함된 형태로 저장됨 — submitToGithub.ts)
-export function formatIssueKey(issue: Pick<IssueRecord, "platform" | "key">): string {
-  if (!issue.key) return "";
-  return issue.key;
-}
-
-function issueTimestamp(issue: IssueRecord): number {
-  return issue.submittedAt ?? issue.createdAt;
-}
-
-function dateLabel(ts: number): string {
-  const locale = dateBcp47();
-  return new Date(ts).toLocaleDateString(locale, {
-    year: "numeric",
-    month: locale === "ko-KR" ? "long" : "short",
-    day: "numeric",
-  });
-}
-
-function formatDate(ts: number, t: TranslationFn): string {
-  const now = Date.now();
-  const diff = now - ts;
-  if (diff < 60_000) return t("time.justNow");
-  if (diff < 3_600_000) return t("time.minutesAgo", { n: Math.floor(diff / 60_000) });
-  if (diff < 86_400_000) return t("time.hoursAgo", { n: Math.floor(diff / 3_600_000) });
-  if (diff < 7 * 86_400_000) return t("time.daysAgo", { n: Math.floor(diff / 86_400_000) });
-  return dateLabel(ts);
-}
-
-function groupByDate(issues: IssueRecord[]): [string, IssueRecord[]][] {
-  const groups = new Map<string, IssueRecord[]>();
-  for (const issue of issues) {
-    const key = dateLabel(issueTimestamp(issue));
-    const list = groups.get(key);
-    if (list) list.push(issue);
-    else groups.set(key, [issue]);
-  }
-  return Array.from(groups);
 }
