+++ src/sidepanel/tabs/jiraFields/AssigneeField.tsx
@@ -1,0 +1,87 @@
+import { useCallback, useEffect, useState } from "react";
+import { Check } from "lucide-react";
+import { useT } from "@/i18n";
+import { CommandItem } from "@/components/ui/command";
+import { cn } from "@/lib/utils";
+import type { JiraUser } from "@/types/jira";
+import { sendBg } from "@/types/messages";
+import { FieldCombobox } from "./FieldCombobox";
+import { useDebouncedSearch } from "./useDebouncedSearch";
+import { useJiraConfig } from "./useJiraConfig";
+
+export function AssigneeField({
+  value,
+  fallbackLabel,
+  onChange,
+}: {
+  value?: string;
+  fallbackLabel?: string;
+  onChange: (id: string | undefined, name?: string) => void;
+}) {
+  const t = useT();
+  const jira = useJiraConfig();
+  const [open, setOpen] = useState(false);
+
+  const fetchUsers = useCallback(
+    (query: string) => {
+      if (!jira) return Promise.resolve([]);
+      return sendBg<JiraUser[]>({
+        type: "jira.searchUsers",
+        query,
+      });
+    },
+    [jira],
+  );
+
+  const { items, loading, error, search } = useDebouncedSearch(fetchUsers);
+
+  useEffect(() => {
+    if (open) return search("");
+  }, [open, search]);
+
+  const selected = items.find((u) => u.accountId === value);
+
+  return (
+    <FieldCombobox
+      open={open}
+      onOpenChange={setOpen}
+      loading={loading}
+      error={error}
+      placeholder={t("field.assignee.select")}
+      searchPlaceholder={t("field.assignee.search")}
+      emptyMessage={t("field.assignee.empty")}
+      label={selected?.displayName}
+      fallbackLabel={fallbackLabel}
+      clearable={!!value}
+      onClear={() => onChange(undefined)}
+      onSearch={search}
+      groupLabel={t("field.assignee.label")}
+    >
+      {items.map((u) => (
+        <CommandItem
+          key={u.accountId}
+          value={u.displayName}
+          onSelect={() => {
+            onChange(u.accountId, u.displayName);
+            setOpen(false);
+          }}
+        >
+          <Check
+            className={cn(
+              "mr-2 h-4 w-4",
+              value === u.accountId ? "opacity-100" : "opacity-0",
+            )}
+          />
+          {u.avatarUrls?.["16x16"] ? (
+            <img
+              src={u.avatarUrls["16x16"]}
+              alt=""
+              className="mr-2 h-4 w-4 rounded-full"
+            />
+          ) : null}
+          <span className="min-w-0 flex-1 truncate">{u.displayName}</span>
+        </CommandItem>
+      ))}
+    </FieldCombobox>
+  );
+}
