+++ src/sidepanel/tabs/jiraFields/JiraIssueFields.tsx
@@ -1,0 +1,56 @@
+import { useT } from "@/i18n";
+import { FieldRow } from "@/sidepanel/components/FieldRow";
+import type { EditorIssueFields } from "@/store/editor-store";
+import { AssigneeField } from "./AssigneeField";
+import { EpicField } from "./EpicField";
+import { IssueTypeField } from "./IssueTypeField";
+import { PriorityField } from "./PriorityField";
+
+export function JiraIssueFields({
+  fields,
+  onChange,
+}: {
+  fields: EditorIssueFields;
+  onChange: (patch: Partial<EditorIssueFields>) => void;
+}) {
+  const t = useT();
+  return (
+    <div className="flex flex-col gap-4">
+      <FieldRow label={t("create.issueType")} required>
+        <IssueTypeField
+          value={fields.issueTypeId}
+          onChange={(id) => onChange({ issueTypeId: id })}
+        />
+      </FieldRow>
+      <FieldRow label={t("create.assignee")}>
+        <AssigneeField
+          value={fields.assigneeId}
+          fallbackLabel={fields.assigneeName}
+          onChange={(id, name) => onChange({ assigneeId: id, assigneeName: name })}
+        />
+      </FieldRow>
+      <FieldRow label={t("create.priority")}>
+        <PriorityField
+          value={fields.priorityId}
+          fallbackLabel={fields.priorityName}
+          onChange={(id, name) => onChange({ priorityId: id, priorityName: name })}
+        />
+      </FieldRow>
+      <FieldRow label={t("create.parentEpic")}>
+        <EpicField
+          value={fields.parentKey}
+          fallbackLabel={fields.parentLabel}
+          onChange={(key, label) => onChange({ parentKey: key, parentLabel: label })}
+          hierarchyLevels={[1]}
+        />
+      </FieldRow>
+      <FieldRow label={t("create.linkedIssue")}>
+        <EpicField
+          value={fields.relatesKey}
+          fallbackLabel={fields.relatesLabel}
+          onChange={(key, label) => onChange({ relatesKey: key, relatesLabel: label })}
+        />
+      </FieldRow>
+    </div>
+  );
+}
