+++ src/sidepanel/tabs/linearFields/LinearIssueFields.tsx
@@ -1,6 +1,6 @@
 import { useT } from "@/i18n";
 import type { LinearDefaults } from "@/types/linear";
-import { FieldRow } from "@/sidepanel/tabs/IssueCreateModal";
+import { FieldRow } from "@/sidepanel/components/FieldRow";
 import { AssigneeCombobox } from "./AssigneeCombobox";
 import { LabelCombobox } from "./LabelCombobox";
 import { PrioritySelect } from "./PrioritySelect";
